chore: Reorganize component tests so that ones needing browsers are within test/system_components and end with _system_test.rb
