+++ .github/workflows/ci.yml
@@ -176,7 +176,7 @@ jobs:
       - name: Run tests
         env:
           RAILS_ENV: test
-          DEFAULT_TEST_EXCLUDE: test/{system,components,integration/sapporo,dummy,fixtures}/**/*_test.rb
+          DEFAULT_TEST_EXCLUDE: test/{system,integration/sapporo,dummy,fixtures}/**/*_test.rb
         run: bin/rails db:test:prepare test
 
   system-test:
@@ -246,7 +246,7 @@ jobs:
       - name: Run System Tests
         env:
           RAILS_ENV: test
-        run: bin/rails db:test:prepare test:system test:components
+        run: bin/rails db:test:prepare test:system
 
       - name: Keep screenshots from failed system tests
         uses: actions/upload-artifact@v7
