--- test/components/system/form_error_summary_component_test.rb
@@ -1,36 +1,0 @@
-# frozen_string_literal: true
-
-require 'application_system_test_case'
-
-module System
-  class FormErrorSummaryComponentTest < ApplicationSystemTestCase
-    test 'focus demo' do
-      visit('rails/view_components/form_error_summary_component/focus_demo')
-
-      assert_selector '[data-controller="form-error-summary"]', focused: true
-      assert_selector 'div[data-controller="form-error-summary"][tabindex="-1"]'
-
-      within '[data-controller="form-error-summary"]' do
-        click_link "Email can't be blank"
-      end
-      assert_selector '#user_email', focused: true
-
-      within '[data-controller="form-error-summary"]' do
-        click_link 'Namespace required'
-      end
-      assert_selector '#namespace-select', focused: true
-
-      within '[data-controller="form-error-summary"]' do
-        click_link "Expiration date can't be blank"
-      end
-      assert_selector '#expires_at-input', focused: true
-
-      within '[data-controller="form-error-summary"]' do
-        click_link "Scopes can't be blank"
-      end
-      assert_selector '#personal_access_token_scopes_api', focused: true
-
-      assert_accessible
-    end
-  end
-end
