--- test/components/viral/pagy_full_component_preview_test.rb
@@ -1,24 +1,0 @@
-# frozen_string_literal: true
-
-require 'application_system_test_case'
-
-class PagyFullComponentPreviewTest < ApplicationSystemTestCase
-  test 'renders default' do
-    visit('/rails/view_components/viral_pagy_full_component/default')
-
-    assert_selector 'nav.pagy.nav'
-    assert_selector 'li span.cursor-not-allowed', text: I18n.t('components.viral.pagy.pagination_component.previous')
-    assert_selector 'li > a', text: I18n.t('components.viral.pagy.pagination_component.next')
-    assert_selector 'li span.cursor-not-allowed', text: I18n.t('components.viral.pagy.pagination_component.previous')
-    assert_selector 'li a[aria-current="page"]', text: '1', count: 1
-    assert_selector 'li > a', count: '6'
-  end
-
-  test 'renders empty state' do
-    visit('/rails/view_components/viral_pagy_full_component/empty_state')
-
-    assert_selector 'h2', text: I18n.t('components.viral.pagy.empty_state.title')
-    assert_selector 'span', text: I18n.t('components.viral.pagy.empty_state.description')
-    assert_no_selector 'nav.pagy.nav'
-  end
-end
