+++ test/components/viral/pagy_pagination_component_test.rb
